+++ packages/liveblocks-react/src/types/index.ts
@@ -17,7 +17,6 @@ import type {
   AiChat,
   AiChatMessage,
   AiChatsQuery,
-  AiKnowledgeSource,
   AiUserMessage,
   AsyncError,
   AsyncLoading,
@@ -89,11 +88,6 @@ export type UseSendAiMessageOptions = {
    * The maximum timeout for the answer to be generated.
    */
   timeout?: number;
-
-  /**
-   * @internal
-   */
-  knowledge?: AiKnowledgeSource[];
 };
 
 export type SendAiMessageOptions = UseSendAiMessageOptions & {