+++ CHANGELOG.md
@@ -1,9 +1,30 @@
+## vNEXT (not yet published)
+
+## v3.8.1
+
+### `@liveblocks/react`
+
+- Add `chatId` prop to `RegisterAiKnowledge` to scope knowledge to a specific
+  chat, similar to `RegisterAiTool`. This is the same as using the `knowledge`
+  prop on `AiChat`.
+- Fix issue where `useAiChat()` didn't re-render correctly when chat title gets
+  updated.
+
+### `@liveblocks/node`
+
+- Fix issue where `tenantId` was not being passed to the request when using
+  `Liveblocks.createRoom()`.
+
+- Add `comments:write` to the list of possible room permissions.
+
 ## v3.8.0
 
 ### `@liveblocks/client`
 
 - LiveMap and LiveObject deletions now report which item got deleted in the
   update notifications. LiveLists already did this.
+- Support numerical operators `gt`, `lt`, `gte`, and `lte` in `room.getThreads`
+  metadata query filters.
 
 ### `@liveblocks/react`
 
@@ -14,13 +35,7 @@
   if so, what type of content is currently generating.
 - Fixes an issue where `useUnreadInboxNotificationsCount` wasn't returning the
   proper count if there were more than a page of unread notifications.
-
 - Support numerical operators `gt`, `lt`, `gte`, and `lte` in `useThreads`
-  metadata query filters.
-
-### `@liveblocks/client`
-
-- Support numerical operators `gt`, `lt`, `gte`, and `lte` in `room.getThreads`
   metadata query filters.
 
 ### `@liveblocks/react-ui`